+++ lib/src/board_datetime_builder.dart
@@ -191,7 +191,7 @@ class _BoardDateTimeBuilderState<T extends BoardDateTimeCommonResult>
 
     Widget child() {
       return SingleBoardDateTimeContent<T>(
-        key: widget.controller._key,
+        key: widget.controller.boardKey,
         onChange: widget.onChange,
         onResult: widget.onResult,
         pickerType: widget.pickerType,
@@ -253,15 +253,12 @@ class SingleBoardDateTimeContent<T extends BoardDateTimeCommonResult>
     super.onKeyboadClose,
     super.onUpdateByClose,
     required super.headerWidget,
-    required this.onTopActionBuilder,
+    required super.onTopActionBuilder,
     super.customCloseButtonBuilder,
   });
 
   final void Function(DateTime)? onChange;
   final void Function(T)? onResult;
-
-  /// Specify a Widget to be displayed in the action button area externally
-  final Widget Function(BuildContext context)? onTopActionBuilder;
 
   final DateTime? initialDate;
 
@@ -289,9 +286,7 @@ class _SingleBoardDateTimeContentState<T extends BoardDateTimeCommonResult>
     super.dispose();
   }
 
-  @override
-  void setNewValue(DateTime val, {bool byPicker = false}) {
-    dateState.value = val;
+  void _setFocusNode(bool byPicker) {
     if (byPicker && widget.pickerFocusNode != null) {
       final fn = widget.pickerFocusNode!;
       if (!fn.hasFocus &&
@@ -299,6 +294,12 @@ class _SingleBoardDateTimeContentState<T extends BoardDateTimeCommonResult>
         fn.requestFocus();
       }
     }
+  }
+
+  @override
+  void setNewValue(DateTime val, {bool byPicker = false}) {
+    dateState.value = val;
+    _setFocusNode(byPicker);
   }
 
   @override
@@ -326,6 +327,18 @@ class _SingleBoardDateTimeContentState<T extends BoardDateTimeCommonResult>
       BoardDateTimeCommonResult.init(pickerType, dateState.value) as T,
     );
     changedDate = true;
+  }
+
+  /// Reset date.
+  /// During this process, re-register the Listener to avoid sending unnecessary notifications.
+  void reset() {
+    dateState.removeListener(notify);
+    dateState.value = defaultDate ?? DateTime.now();
+    changeDateTime(dateState.value);
+    dateState.addListener(notify);
+
+    notify();
+    _setFocusNode(false);
   }
 
   @override
@@ -360,7 +373,7 @@ class _SingleBoardDateTimeContentState<T extends BoardDateTimeCommonResult>
       maximumDate: widget.maximumDate,
       multiple: false,
       headerBuilder: (ctx) => _header,
-      onChange: changeDate,
+      onChangeByCalendar: changeDate,
       onChangeByPicker: onChangeByPicker,
       onKeyboadClose: closeKeyboard,
       keyboardHeightRatio: () => keyboardHeightRatio,
@@ -450,6 +463,7 @@ class _SingleBoardDateTimeContentState<T extends BoardDateTimeCommonResult>
       topMargin: widget.options.topMargin,
       onTopActionBuilder: widget.onTopActionBuilder,
       actionButtonTypes: widget.options.actionButtonTypes,
+      onReset: widget.options.useResetButton ? reset : null,
       customCloseButtonBuilder: widget.customCloseButtonBuilder,
     );
   }