+++ lib/src/board_datetime_widget.dart
@@ -415,6 +415,7 @@ class _SingleBoardDateTimeWidgetState
 Future<BoardDateTimeMultiSelection?>
     showBoardDateTimeMultiPicker<T extends BoardDateTimeCommonResult>({
   required BuildContext context,
+  BoardMultiDateTimeController? controller,
   required DateTimePickerType pickerType,
   // ValueNotifier<DateTime>? valueNotifier,
   void Function(BoardDateTimeMultiSelection)? onChanged,
@@ -435,6 +436,7 @@ Future<BoardDateTimeMultiSelection?>
   bool enableDrag = true,
   bool? showDragHandle,
   bool useSafeArea = false,
+  Widget Function(BuildContext context)? onTopActionBuilder,
 }) async {
   final opt = options ?? const BoardDateTimeOptions();
 
@@ -464,6 +466,7 @@ Future<BoardDateTimeMultiSelection?>
           bottom: MediaQuery.of(ctx).viewInsets.bottom,
         ),
         child: _MultiBoardDateTimeWidget(
+          controller: controller,
           breakpoint: breakpoint,
           pickerType: pickerType,
           startDate: startDate,
@@ -475,6 +478,7 @@ Future<BoardDateTimeMultiSelection?>
           // valueNotifier: valueNotifier,
           onChanged: onChanged,
           onResult: (val1, val2) => onResult?.call(val1 as T, val2 as T),
+          onTopActionBuilder: onTopActionBuilder,
         ),
       );
     },
@@ -483,6 +487,7 @@ Future<BoardDateTimeMultiSelection?>
 
 class _MultiBoardDateTimeWidget extends StatefulWidget {
   const _MultiBoardDateTimeWidget({
+    this.controller,
     this.breakpoint = 800,
     this.startDate,
     this.endDate,
@@ -494,9 +499,11 @@ class _MultiBoardDateTimeWidget extends StatefulWidget {
     this.onChanged,
     this.onResult,
     this.headerWidget,
+    this.onTopActionBuilder,
     this.customCloseButtonBuilder,
   });
 
+  final BoardMultiDateTimeController? controller;
   final double breakpoint;
   final DateTime? startDate;
   final DateTime? endDate;
@@ -509,6 +516,7 @@ class _MultiBoardDateTimeWidget extends StatefulWidget {
   final void Function(BoardDateTimeMultiSelection)? onChanged;
   final void Function(BoardDateTimeCommonResult, BoardDateTimeCommonResult)?
       onResult;
+  final Widget Function(BuildContext context)? onTopActionBuilder;
   final CloseButtonBuilder? customCloseButtonBuilder;
 
   @override
@@ -521,16 +529,36 @@ class _MultiBoardDateTimeWidgetState extends State<_MultiBoardDateTimeWidget> {
 
   @override
   void initState() {
-    selection = BoardDateTimeMultiSelection(
-      start: widget.startDate ?? DateTime.now(),
-      end: widget.endDate ?? DateTime.now().add(const Duration(days: 1)),
-    );
+    if (widget.pickerType == DateTimePickerType.time) {
+      final now = DateTime.now();
+      selection = BoardDateTimeMultiSelection(
+        start:
+            widget.startDate ?? DateTime(now.year, now.month, now.day, 0, 0, 0),
+        end: widget.endDate ??
+            DateTime(now.year, now.month, now.day, 23, 59, 59),
+      );
+    } else {
+      selection = BoardDateTimeMultiSelection(
+        start: widget.startDate ?? DateTime.now(),
+        end: widget.endDate ?? DateTime.now().add(const Duration(days: 1)),
+      );
+    }
     super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
+    DateTime? minimumDate = widget.minimumDate;
+    DateTime? maximumDate = widget.maximumDate;
+
+    if (widget.pickerType == DateTimePickerType.time) {
+      final now = DateTime.now();
+      minimumDate ??= DateTime(now.year, now.month, now.day, 0, 0, 0);
+      maximumDate ??= DateTime(now.year, now.month, now.day, 23, 59, 59);
+    }
+
     return MultiBoardDateTimeContent(
+      key: widget.controller?.boardKey,
       onChange: (start, end) {
         selection = BoardDateTimeMultiSelection(start: start, end: end);
         // widget.valueNotifier?.value = val;
@@ -542,8 +570,8 @@ class _MultiBoardDateTimeWidgetState extends State<_MultiBoardDateTimeWidget> {
       breakpoint: widget.breakpoint,
       startDate: selection.start,
       endDate: selection.end,
-      minimumDate: widget.minimumDate,
-      maximumDate: widget.maximumDate,
+      minimumDate: minimumDate,
+      maximumDate: maximumDate,
       headerWidget: widget.headerWidget,
       modal: true,
       onCloseModal: () {
@@ -555,6 +583,7 @@ class _MultiBoardDateTimeWidgetState extends State<_MultiBoardDateTimeWidget> {
           end: val2 ?? selection.end,
         );
       },
+      onTopActionBuilder: widget.onTopActionBuilder,
       customCloseButtonBuilder: widget.customCloseButtonBuilder,
     );
   }