+++ lib/src/ui/board_datetime_contents_state.dart
@@ -1,5 +1,6 @@
 import 'dart:math';
 
+import 'package:board_datetime_picker/src/ui/parts/header.dart';
 import 'package:board_datetime_picker/src/utils/board_datetime_options_extension.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/date_symbol_data_local.dart';
@@ -28,6 +29,7 @@ abstract class BoardDateTimeContent<T extends BoardDateTimeCommonResult>
     this.onKeyboadClose,
     this.onUpdateByClose,
     required this.headerWidget,
+    required this.onTopActionBuilder,
     required this.customCloseButtonBuilder,
   });
 
@@ -61,11 +63,10 @@ abstract class BoardDateTimeContent<T extends BoardDateTimeCommonResult>
   /// To be displayed at the top of the picker
   final Widget? headerWidget;
 
-  final Widget Function(
-    BuildContext context,
-    bool isModal,
-    void Function() onClose,
-  )? customCloseButtonBuilder;
+  /// Specify a Widget to be displayed in the action button area externally
+  final Widget Function(BuildContext context)? onTopActionBuilder;
+
+  final CloseButtonBuilder? customCloseButtonBuilder;
 }
 
 abstract class BoardDatetimeContentState<T extends BoardDateTimeCommonResult,
@@ -265,6 +266,7 @@ abstract class BoardDatetimeContentState<T extends BoardDateTimeCommonResult,
           final subTitle = widget.options.getSubTitle(DateType.year);
           ymdOptions.add(
             initItemOption(
+              widget.pickerType,
               DateType.year,
               d,
               minDate,
@@ -272,12 +274,14 @@ abstract class BoardDatetimeContentState<T extends BoardDateTimeCommonResult,
               null,
               subTitle,
               withSecond,
+              false,
             ),
           );
         } else if (pf == 'M') {
           final subTitle = widget.options.getSubTitle(DateType.month);
           ymdOptions.add(
             initItemOption(
+              widget.pickerType,
               DateType.month,
               d,
               minDate,
@@ -285,12 +289,14 @@ abstract class BoardDatetimeContentState<T extends BoardDateTimeCommonResult,
               null,
               subTitle,
               withSecond,
+              false,
             ),
           );
         } else if (pf == 'd') {
           final subTitle = widget.options.getSubTitle(DateType.day);
           ymdOptions.add(
             initItemOption(
+              widget.pickerType,
               DateType.day,
               d,
               minDate,
@@ -298,6 +304,7 @@ abstract class BoardDatetimeContentState<T extends BoardDateTimeCommonResult,
               null,
               subTitle,
               withSecond,
+              false,
             ),
           );
         }
@@ -310,6 +317,7 @@ abstract class BoardDatetimeContentState<T extends BoardDateTimeCommonResult,
       if ([DateTimePickerType.time, DateTimePickerType.datetime]
           .contains(type)) ...[
         initItemOption(
+          widget.pickerType,
           DateType.hour,
           d,
           minDate,
@@ -317,8 +325,10 @@ abstract class BoardDatetimeContentState<T extends BoardDateTimeCommonResult,
           opts?.hours,
           widget.options.getSubTitle(DateType.hour),
           withSecond,
+          widget.options.useAmpm && type == DateTimePickerType.time,
         ),
         initItemOption(
+          widget.pickerType,
           DateType.minute,
           d,
           minDate,
@@ -326,10 +336,12 @@ abstract class BoardDatetimeContentState<T extends BoardDateTimeCommonResult,
           opts?.minutes,
           widget.options.getSubTitle(DateType.minute),
           withSecond,
+          false,
         ),
       ],
       if (DateTimePickerType.time == type && widget.options.withSecond)
         initItemOption(
+          widget.pickerType,
           DateType.second,
           d,
           minDate,
@@ -337,6 +349,7 @@ abstract class BoardDatetimeContentState<T extends BoardDateTimeCommonResult,
           opts?.seconds,
           widget.options.getSubTitle(DateType.second),
           withSecond,
+          false,
         ),
     ];
 
@@ -363,7 +376,6 @@ abstract class BoardDatetimeContentState<T extends BoardDateTimeCommonResult,
   void onChangeByPicker(BoardPickerItemOption opt, int index) {
     // Update option class values to selected values
     opt.selectedIndex = index;
-
     // 年と月から選択中の日付が存在するか確認する
     // Check to see if the date you are selecting exists from the year and month
     DateTime newVal;
@@ -383,7 +395,7 @@ abstract class BoardDatetimeContentState<T extends BoardDateTimeCommonResult,
       newVal = opt.calcDate(currentDate);
     }
 
-    final data = opt.map[index]!;
+    final data = opt.itemMap[index]!;
     final day = DateTimeUtil.getExistsMaxDate(itemOptions, opt, data);
     if (day != null) {
       final dayOpt = itemOptions.firstWhere((x) => x.type == DateType.day);
@@ -449,7 +461,7 @@ abstract class BoardDatetimeContentState<T extends BoardDateTimeCommonResult,
     setNewValue(newVal);
   }
 
-  void changeDateTime(DateTime val) {
+  void changeDateTime(DateTime val, {bool needNotify = true}) {
     DateTime newVal = DateTimeUtil.rangeDate(
       val,
       minimumDate,
@@ -471,6 +483,8 @@ abstract class BoardDatetimeContentState<T extends BoardDateTimeCommonResult,
         x.changeDate(newVal);
       }
     }
-    setNewValue(newVal);
+    if (needNotify) {
+      setNewValue(newVal);
+    }
   }
 }